# Ix: Implement the Retry() operator directly