--- a/Ix.NET/Source/System.Interactive/Retry.cs
+++ b/Ix.NET/Source/System.Interactive/Retry.cs
@@ -82,8 +82,6 @@ namespace System.Linq
 
                         yield return v;
                     }
-
-                    continue;
                 }
             }
         }
@@ -127,8 +125,6 @@ namespace System.Linq
 
                         yield return v;
                     }
-
-                    continue;
                 }
             }