This PR reimplements the Ix `Retry()` operators (infinite, max count) directly.

(The comparison benchmark will be added next morning.)