@akarnokd 2018-07-04
```
BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i7-4790 CPU 3.60GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
Frequency=3513585 Hz, Resolution=284.6096 ns, Timer=TSC
.NET Core SDK=2.1.301
  [Host]     : .NET Core 2.1.1 (CoreCLR 4.6.26606.02, CoreFX 4.6.26606.05), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.1 (CoreCLR 4.6.26606.02, CoreFX 4.6.26606.05), 64bit RyuJIT
```

#### Before

   Method |       N |            Mean |           Error |         StdDev |  Gen 0 | Allocated |
--------- |--------:|----------------:|----------------:|---------------:|-------:|----------:|
   Finite |       1 |        163.4 ns |       0.5941 ns |      0.5267 ns | 0.0741 |     312 B |
 Infinite |       1 |        158.5 ns |       0.4885 ns |      0.3814 ns | 0.0703 |     296 B |
   Finite |      10 |        278.6 ns |       1.0634 ns |      0.9947 ns | 0.0739 |     312 B |
 Infinite |      10 |        271.6 ns |       1.0752 ns |      1.0057 ns | 0.0701 |     296 B |
   Finite |     100 |      1,306.6 ns |       7.8899 ns |      7.3802 ns | 0.0725 |     312 B |
 Infinite |     100 |      1,283.1 ns |       5.9046 ns |      5.5232 ns | 0.0687 |     296 B |
   Finite |    1000 |     11,526.5 ns |      68.7482 ns |     53.6741 ns | 0.0610 |     312 B |
 Infinite |    1000 |     11,511.6 ns |      49.3124 ns |     46.1269 ns | 0.0610 |     296 B |
   Finite |   10000 |    113,283.1 ns |     159.8151 ns |    133.4529 ns |      - |     312 B |
 Infinite |   10000 |    112,771.0 ns |     294.2639 ns |    245.7237 ns |      - |     296 B |
   Finite |  100000 |  1,130,615.5 ns |   4,486.8399 ns |  3,977.4664 ns |      - |     312 B |
 Infinite |  100000 |  1,133,064.5 ns |   4,083.3295 ns |  3,819.5486 ns |      - |     296 B |
   Finite | 1000000 | 11,329,297.2 ns | 109,025.0436 ns | 96,647.8519 ns |      - |     312 B |
 Infinite | 1000000 | 11,304,929.8 ns |  44,805.6468 ns | 39,719.0349 ns |      - |     296 B |

#### After
 
   Method |       N |             Mean |          Error |         StdDev |  Gen 0 | Allocated |
--------- |--------:|-----------------:|---------------:|---------------:|-------:|----------:|
   Finite |       1 |         62.10 ns |      0.3173 ns |      0.2968 ns | 0.0418 |     176 B |
 Infinite |       1 |         69.01 ns |      0.2948 ns |      0.2613 ns | 0.0401 |     168 B |
   Finite |      10 |        182.16 ns |      0.5943 ns |      0.5559 ns | 0.0417 |     176 B |
 Infinite |      10 |        175.17 ns |      0.4377 ns |      0.4095 ns | 0.0401 |     168 B |
   Finite |     100 |      1,258.18 ns |     10.0392 ns |      9.3907 ns | 0.0401 |     176 B |
 Infinite |     100 |      1,217.75 ns |      4.6318 ns |      4.3326 ns | 0.0401 |     168 B |
   Finite |    1000 |     12,024.44 ns |     91.3169 ns |     80.9500 ns | 0.0305 |     176 B |
 Infinite |    1000 |     11,658.27 ns |     68.4315 ns |     64.0108 ns | 0.0305 |     168 B |
   Finite |   10000 |    119,027.20 ns |    881.4477 ns |    824.5066 ns |      - |     176 B |
 Infinite |   10000 |    115,026.97 ns |    244.3036 ns |    228.5217 ns |      - |     168 B |
   Finite |  100000 |  1,183,478.55 ns |  5,195.5034 ns |  4,859.8766 ns |      - |     176 B |
 Infinite |  100000 |  1,140,880.58 ns |  7,090.8103 ns |  6,632.7477 ns |      - |     168 B |
   Finite | 1000000 | 11,780,148.68 ns | 39,165.5683 ns | 34,719.2527 ns |      - |     176 B |
 Infinite | 1000000 | 11,365,741.83 ns | 88,983.5910 ns | 83,235.2982 ns |      - |     168 B |