@akarnokd 2018-07-04
About the benchmark: for shorter sequences, it works better but longer sequences don't get any additional benefit. The allocation amount seem sto be reduced though.